Couple changes in the Room and Peer classes.
