+++ server/src/main/java/rooms/Peer.java
@@ -1,4 +1,4 @@
-package server;
+package rooms;
 
 import utils.Address;
